% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-raster.R
\name{adjust_pop_by_shp}
\alias{adjust_pop_by_shp}
\title{Adjust population raster to match values from shapefile}
\usage{
adjust_pop_by_shp(input_raster, ref_shp, col_name, output_class = "terra")
}
\arguments{
\item{input_raster}{A SpatRaster (or RasterLayer) objects that will be adjusted such that the areal sums match the target column in ref_shp.}

\item{ref_shp}{A sf polygon object.}

\item{col_name}{A character string the identifies which column in ref_shp will be used to scale input_raster (this column must have numerical values and should not contain NAs).}

\item{output_class}{Either "terra" (default) or "raster".}
}
\value{
a SpatRaster (or RasterLayer if output_class == "raster") object
}
\description{
Adjust population raster to match values from shapefile
}
