% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-shapefile.R
\name{available_shapefile}
\alias{available_shapefile}
\title{Check if shapefile is available}
\usage{
available_shapefile(country = NULL, admin_level = NULL)
}
\arguments{
\item{country}{Name of country}

\item{admin_level}{Numeric value either 0, 1, or 2.}
}
\value{
Printed message in console
}
\description{
Check if shapefile is available
}
