% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-functions.R
\name{cmc_to_month}
\alias{cmc_to_month}
\title{Convert Century-Month Code date to standard format}
\usage{
cmc_to_month(cmc)
}
\arguments{
\item{cmc}{A Century-Month Code formated date}
}
\value{
A date in standard format
}
\description{
Convert Century-Month Code date to standard format
}
