% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-functions.R
\name{compare_unique_entries}
\alias{compare_unique_entries}
\title{Compare the unique values in two character vectors}
\usage{
compare_unique_entries(vec_a, vec_b, fill_value = NA_character_)
}
\arguments{
\item{vec_a}{A character vector}

\item{vec_b}{A character vector}

\item{fill_value}{A character value, default is NA_character_}
}
\value{
A dataframe
}
\description{
Compare the unique values in two character vectors
}
