% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{define_urban}
\alias{define_urban}
\title{Define urban-only pixels.}
\usage{
define_urban(
  population_raster,
  rururb_cutoff = 300,
  min_urbsize = 2000,
  directions = 8,
  mask = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{population_raster}{An input raster containing people per pixel. Default inputs assume input resolution to be approximately 1 sq. km. resolution.}

\item{rururb_cutoff}{Minimum population per pixel to be eligible for urban classification.}

\item{min_urbsize}{Minimum total population of clustered pixels to the classified as urban.}

\item{directions}{See terra::patches, default is 8.}

\item{mask}{TRUE/FALSE, Should the returned raster have masked values (1 = Urban, 0 = Rural). Default is FALSE.}

\item{verbose}{TRUE/FALSE Print messages?}
}
\value{
A SpatRaster object.
}
\description{
Define urban-only pixels.
}
