% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-functions.R
\name{ethiopian_to_greg}
\alias{ethiopian_to_greg}
\title{Convert Ethiopian date to Gregorian date}
\usage{
ethiopian_to_greg(cmc)
}
\arguments{
\item{cmc}{An Ethiopian formatted date in Century-Month Code system.}
}
\value{
A date in Gregorian format
}
\description{
Convert Ethiopian date to Gregorian date
}
