% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-times.R
\name{get_friction_surface}
\alias{get_friction_surface}
\title{Download friction surface from Malaria Atlas Project}
\usage{
get_friction_surface(
  shp = NULL,
  extent = NULL,
  transport_type = "walking",
  export = FALSE,
  output_dir = "."
)
}
\arguments{
\item{shp}{Area of interest, can be sf of SpatialPolygonDataFrame object, or character string for specific country (will use malariaAtlas shapefile).}

\item{extent}{2x2 matrix specifying the spatial extent within which raster data is desired, as returned by sp::bbox() - the first column has the minimum, the second the maximum values; rows 1 & 2 represent the x & y dimensions respectively (matrix(c("xmin", "ymin","xmax", "ymax"), nrow = 2, ncol = 2, dimnames = list(c("x", "y"), c("min", "max")))) (use either shp OR extent; if neither is specified global raster is returned).}

\item{transport_type}{Either "walking" (default) or "motorized".}

\item{export}{TRUE/FALSE Should output be saved?}

\item{output_dir}{String to output folder location.}
}
\description{
Download friction surface from Malaria Atlas Project
}
