% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_path.R
\name{here_path}
\alias{here_path}
\title{here_path}
\usage{
here_path(..., ext = NULL)
}
\arguments{
\item{...}{passed to [here::here()]}

\item{ext}{file extension}
}
\value{

}
\description{
Returns a path relative to `here::here()`, and creates it if it does not yet exist.
}
\details{
Also applies `glue::glue()` to the path (after `here::here()`).
}
