% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-shapefile.R
\name{load_shapefile}
\alias{load_shapefile}
\title{Functions for loading and downloading shapefiles from
https://github.com/PATH-Global-Health/geometries}
\usage{
load_shapefile(
  country,
  admin_level = c(0, 1, 2),
  quiet = T,
  download = F,
  destfile = "."
)
}
\arguments{
\item{country}{Name of country}

\item{admin_level}{Numeric value either 0, 1, or 2.}

\item{quiet}{sf parameter to choose if shapefile info is printed}

\item{download}{TRUE/FALSE should the GeoJSON be downloaded?}

\item{destfile}{Path to output directory}
}
\value{
A sf-object
}
\description{
Load and download shapefile
}
