% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-functions.R
\name{month_to_cmc}
\alias{month_to_cmc}
\title{Convert data into Century-Month Code system}
\usage{
month_to_cmc(date)
}
\arguments{
\item{date}{A date value}
}
\value{
A date in Century-Month Code format
}
\description{
Convert data into Century-Month Code system
}
