% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-palette.R
\name{path_palette}
\alias{path_palette}
\title{Generating PATH color palettes based on user inputs}
\usage{
path_palette(name = "path_style", n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{name of specific path palette, default is path_style, secondary option for tableau_style}

\item{n}{number of colors, if left undefined it will default to the length of the palette}

\item{type}{continuous or discrete}
}
\description{
Generating PATH color palettes based on user inputs
}
