% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-palette.R
\name{scale_color_path_d}
\alias{scale_color_path_d}
\alias{scale_fill_path_d}
\alias{scale_color_path_c}
\alias{scale_fill_path_c}
\title{PATH color palette functions for ggplot}
\usage{
scale_color_path_d(name = "path_style")

scale_fill_path_d(name = "path_style")

scale_color_path_c(name = "path_style")

scale_fill_path_c(name = "path_style")
}
\arguments{
\item{name}{name of palette}
}
\description{
Various functions for apply PATH color palettes to ggplot graphics. There are
specific functions for setting the "fill" and "color" aesthetics, and
suffixes for continuous (\verb{_c}) and discrete (\verb{_d}) variables. For example,
\code{scale_color_path_d} would apply the PATH color palette to a discrete
variable in the "color" aesthetic.
}
