% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{daily_rainfall}
\alias{daily_rainfall}
\title{Areal rainfall data}
\usage{
daily_rainfall(dates, shapefile, fun = "sum", long = F, verbose = T)
}
\arguments{
\item{dates}{A character vector containing the start and end date in "YYYY-MM-DD" format.}

\item{shapefile}{A sf object containing the areal aggregation polygons}

\item{fun}{A character string containing the aggregation function, default is "sum".}

\item{long}{TRUE/FALSE Should the output be in "long" format? Default is FALSE.}

\item{verbose}{TRUE/FALSE Print progress messages?}
}
\value{
A tibble containing extracted values along with any columns from 'shapefile'.
}
\description{
Areal rainfall data
}
